+++ tests/integration/test_plugin.py
@@ -6,7 +6,7 @@ from delfino.models.pyproject_toml import PluginConfig
 
 from delfino_core.commands.dependencies_update import run_dependencies_update
 from delfino_core.commands.format import run_black, run_ensure_pre_commit, run_group_format, run_isort, run_pyupgrade
-from delfino_core.commands.lint import run_group_lint, run_pycodestyle, run_pydocstyle, run_pylint, run_ruff
+from delfino_core.commands.lint import run_group_lint, run_pycodestyle, run_pylint, run_ruff
 from delfino_core.commands.pre_commit import run_pre_commit
 from delfino_core.commands.switch_python_version import run_switch_python_version
 from delfino_core.commands.test import (
@@ -58,7 +58,6 @@ class TestPlugin:
             run_mypy,
             run_pre_commit,
             run_pycodestyle,
-            run_pydocstyle,
             run_pylint,
             run_switch_python_version,
             run_pytest,